add Gemfile
to specify what ruby components are needd for the build.
To set up an isolated ruby environment:

    $ rvm --create --ruby-version use 1.9.3@angular
    $ bundle install

The rvm command creates the files .ruby-gemset .ruby-version
in the project directory, specifying what version of ruby is
to be used and the name of the isolated repository for ruby
modules (gems), in this case 'angular'.  By sourcing the
RVM install script in your shell profile, 'cd' is overriden,
so that the proper version of ruby and repository are made
available automatically when entering the directory.

The 'bundle install' command downloads the modules
listed in 'Gemfile', plus dependencies.